libxl: do not require filling backend_domid to remove device
Existing device can be unambiguously identified by (domid, devtype,
devid), do not require backend_domid (or _name) to be filled. This
especially helps detaching a device - which with not filled (or invalid)
backend_domid would seemingly succeed, while in fact the backend
wouldn't be cleaned up.

Try to fill the backend_domid only if not explicitly given (left as 0,
as libxl_device_disk_init() does). This allows the function to still
work outside the toolstack domain - specifically in the backend domain
as part of the xendriverdomain service.

Signed-off-by: Marek Marczykowski-Górecki <marmarek@invisiblethingslab.com>