Additional support in ACPI builder to support SLIC and OEM installs.
In order to use Windows OEM install media, the SLIC table must be passed to a
guest. In addition all the OEM table IDs must match the SLIC or Windows will
think it is invalid and the install ends up unactivated.

NOTE: The DSDT does not get updated. This was the same in the original patch.
Not sure why that is but it is being left that way for now since it works w/o
it.

Ported from qemu-acpi-tables.patch by:
Ross Philipson, philipsonr@ainfosec.com, 05/05/2015

Copied from OpenXT project.

Signed-off-by: Marek Marczykowski-Górecki <marmarek@invisiblethingslab.com>