x86/smp: make cpu_state per-CPU
This will be used for parallel AP bring-up.

This commit changes the direction of CPU_STATE_INIT. It used to be
set by BSP and never consumed by AP. Now it signals that AP got
through assembly part of initialization and waits for BSP to call
notifiers that set up data structures required for further
initialization.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>